product scoring complete
